Various changes - Fixed #ifdef and #ifndef alignment - Moved string.h into string_linux.h and string_darwin.h - Fixed crappy gamepad.h hack - Fixed unsigned int cast instead of just fixing the type - Fixed CMake oddities
