+++ CMakeLists.txt
@@ -109,9 +109,6 @@ endif()
 
 if(APPLE)
     add_subdirectory(libs/epoll)
-endif(APPLE)
-
-if(APPLE)
     include_directories(${Cocoa_INCLUDE_DIRS} ${CoreVideo_INCLUDE_DIRS} ${IOKit_INCLUDE_DIRS})
     include_directories(libs/elf)
 endif(APPLE)
@@ -122,7 +119,7 @@ include_directories(libs/hybris/include)
 
 if(APPLE)
     set(HYBRIS_SOURCES libs/hybris/src/cache.c libs/hybris/src/dlfcn.c libs/hybris/src/hack/hooks.cpp libs/hybris/src/hooks_darwin.c libs/hybris/src/logging.c libs/hybris/src/properties.c libs/hybris/src/hooks_shm.c libs/hybris/src/ctype.c libs/hybris/src/sysconf.c libs/hybris/src/jb/dlfcn.c libs/hybris/src/jb/linker.c libs/hybris/src/jb/linker_environ.c libs/hybris/src/jb/linker_format.c libs/hybris/src/jb/rt.c)
-elseif(NOT APPLE)
+else()
     set(HYBRIS_SOURCES libs/hybris/src/cache.c libs/hybris/src/dlfcn.c libs/hybris/src/hooks.c libs/hybris/src/hooks_shm.c libs/hybris/src/logging.c libs/hybris/src/properties.c libs/hybris/src/strlcpy.c libs/hybris/src/ctype.c libs/hybris/src/sysconf.c libs/hybris/src/jb/dlfcn.c libs/hybris/src/jb/linker.c libs/hybris/src/jb/linker_environ.c libs/hybris/src/jb/linker_format.c libs/hybris/src/jb/rt.c libs/hybris/src/hooks_pthread.c libs/hybris/src/hooks_dirent.c libs/hybris/src/hooks_io.c libs/hybris/src/hooks_net.c)
 endif(APPLE)
 
@@ -132,16 +129,12 @@ set(COMMON_SOURCE_FILES src/common/common.cpp src/common/hook.cpp src/client/app
 
 if(APPLE)
     set(COMMON_SOURCE_FILES ${COMMON_SOURCE_FILES} src/minecraft/string_darwin.cpp src/minecraft/function.h)
-elseif(NOT APPLE)
+else()
     set(COMMON_SOURCE_FILES ${COMMON_SOURCE_FILES} src/minecraft/string_linux.cpp)
+    set(CLIENT_SOURCE_FILES ${CLIENT_SOURCE_FILES} src/client/amdfix.s)
 endif(APPLE)
 
 set(CLIENT_SOURCE_FILES src/client/main.cpp src/client/appplatform.cpp src/xbox/msa.cpp src/xbox/msa.h src/xbox/msa_token.h src/xbox/msa_network.cpp src/xbox/msa_network.h src/xbox/xboxlive.h src/xbox/xboxlive.cpp src/common/cll.cpp src/common/cll.h)
-
-if(NOT APPLE)
-    set(CLIENT_SOURCE_FILES ${CLIENT_SOURCE_FILES} src/client/amdfix.s)
-endif(NOT APPLE)
-
 set(CLIENT_SOURCE_FILES_GAMEPAD src/client/gamepad.cpp src/client/gamepad.h)
 set(CLIENT_SOURCE_FILES_CEF src/common/browser.cpp src/common/browser.h src/xbox/xbox_login_browser.h src/xbox/xbox_login_browser.cpp src/client/initial_setup_browser.cpp src/client/initial_setup_browser.h src/client/window.h)
 set(CLIENT_SOURCE_FILES_CEF_GAMEPAD src/client/gamepad_mapper_browser.cpp src/client/gamepad_mapper_browser.h)
@@ -166,7 +159,7 @@ else()
     include_directories(libs/eglut)
     set(CLIENT_SOURCE_FILES ${CLIENT_SOURCE_FILES} ${CLIENT_SOURCE_FILES_EGLUT} ${EGLUT_SOURCES})
 endif()
-if (GAMEPAD_SUPPORT_ENABLED AND NOT APPLE)
+if (GAMEPAD_SUPPORT_ENABLED)
     set(CLIENT_SOURCE_FILES ${CLIENT_SOURCE_FILES} ${CLIENT_SOURCE_FILES_GAMEPAD})
     if (CEF_ENABLED)
         set(CLIENT_SOURCE_FILES ${CLIENT_SOURCE_FILES} ${CLIENT_SOURCE_FILES_CEF_GAMEPAD})
@@ -194,7 +187,7 @@ if (BUILD_CLIENT)
     else()
         target_compile_definitions(mcpelauncher PRIVATE USE_EGLUT)
     endif()
-    if (GAMEPAD_SUPPORT_ENABLED AND NOT APPLE)
+    if (GAMEPAD_SUPPORT_ENABLED)
         target_link_libraries(mcpelauncher evdev udev)
         target_compile_definitions(mcpelauncher PRIVATE GAMEPAD_SUPPORT)
     endif()
@@ -205,7 +198,7 @@ if (BUILD_CLIENT)
     if(APPLE)
         add_dependencies(mcpelauncher epoll)
         target_link_libraries(mcpelauncher ${CMAKE_DL_LIBS} ${CMAKE_THREAD_LIBS_INIT} ${ZLIB_LIBRARIES} ${LIBZIP_LIBRARIES} ${OPENSSL_LIBRARIES} ${CURL_LIBRARIES} ${Cocoa_LIBRARIES} ${CoreVideo_LIBRARIES} ${IOKit_LIBRARIES} ${OPENGL_LIBRARIES} epoll)
-    elseif(NOT APPLE)
+    else()
         target_link_libraries(mcpelauncher ${CMAKE_DL_LIBS} ${CMAKE_THREAD_LIBS_INIT} ${X11_X11_LIB} ${ZLIB_LIBRARIES} ${PNG_LIBRARIES} ${LIBZIP_LIBRARIES} ${OPENSSL_LIBRARIES} ${CURL_LIBRARIES} rt EGL GLESv2)
     endif(APPLE)
 
@@ -222,7 +215,7 @@ if (BUILD_SERVER)
     if(APPLE)
         add_dependencies(server epoll)
         target_link_libraries(server ${CMAKE_DL_LIBS} ${CMAKE_THREAD_LIBS_INIT} ${ZLIB_LIBRARIES} ${LIBZIP_LIBRARIES} epoll)
-    elseif(NOT APPLE)
+    else()
         target_link_libraries(server ${CMAKE_DL_LIBS} ${CMAKE_THREAD_LIBS_INIT} ${ZLIB_LIBRARIES} ${LIBZIP_LIBRARIES} rt)
     endif(APPLE)
