+++ libs/hybris/src/hooks_shm.c
@@ -94,11 +94,11 @@ static void _sync_mmap_with_shm()
             /* Note that mremap may change the address pointed by _hybris_shm_data.
              * But as we never point directly into _hybris_shm_data, it's fine.
              * */
-             #ifndef __APPLE__
+        #ifndef __APPLE__
              _hybris_shm_data = (hybris_shm_data_t *)mremap( _hybris_shm_data, _current_mapped_size,
                                   _hybris_shm_data->max_offset + HYBRIS_SHM_DATA_HEADER_SIZE,
                                   MREMAP_MAYMOVE );
-             #endif
+        #endif
 
             _current_mapped_size = _hybris_shm_data->max_offset + HYBRIS_SHM_DATA_HEADER_SIZE;
         }
