+++ libs/hybris/src/properties.c
@@ -42,14 +42,12 @@
 #include "properties_p.h"
 
 #ifdef __APPLE__
-#ifndef TEMP_FAILURE_RETRY
 #define TEMP_FAILURE_RETRY(expression) \
-  (__extension__                                                              \
-    ({ long int __result;                                                     \
-       do __result = (long int) (expression);                                 \
-       while (__result == -1L && errno == EINTR);                             \
-       __result; }))
-#endif
+(__extension__                                                              \
+  ({ long int __result;                                                     \
+     do __result = (long int) (expression);                                 \
+     while (__result == -1L && errno == EINTR);                             \
+     __result; }))
 #endif
