+++ src/server/server.cpp
@@ -71,11 +71,11 @@ int main(int argc, char *argv[]) {
     stubSymbols(fmod_symbols, (void*) stubFunc);
     hybris_hook("eglGetProcAddress", (void*) stubFunc);
     hybris_hook("mcpelauncher_hook", (void*) hookFunction);
-    #ifdef __APPLE__
+#ifdef __APPLE__
     void* libmLib = loadLibraryOS("libm.dylib", libm_symbols);
-    #else
+#else
     void* libmLib = loadLibraryOS("libm.so.6", libm_symbols);
-    #endif
+#endif
     hookAndroidLog();
     if (!load_empty_library("libc.so") || !load_empty_library("libm.so"))
         return -1;
